requirements.txt
Flask
gunicorn
psycopg2-binary
google-cloud-secret-manager
google-cloud-storage
fsspec
gcsfs
google-resumable-media
google-auth
Werkzeug
pytz